o Merge with Linux 2.1.91.  o First round of bugfixes for the SC/MC CPUs.  o FPU context switch fixes.  o Lazy context switches.  o Faster syscalls.  o Removed dead code.  o Shitloads of other things I forgot ...

+++ drivers/scsi/wd33c93.c
@@ -1490,7 +1490,7 @@ int wd33c93_abort (Scsi_Cmnd *cmd)
 }
 
 #define MAX_WD33C93_HOSTS 4
-#define MAX_SETUP_ARGS (sizeof(setup_args) / sizeof(char *))
+#define MAX_SETUP_ARGS ((int)(sizeof(setup_args) / sizeof(char *)))
 #define SETUP_BUFFER_SIZE 200
 static char setup_buffer[SETUP_BUFFER_SIZE];
 static char setup_used[MAX_SETUP_ARGS];
